refactor: split type in `src/model/index`
* change `import type` to `import`
