+++ apps/calendar/src/components/timeGrid/timeGrid.tsx
@@ -13,7 +13,6 @@ import { ColumnInfo, ColumnWithMouse } from '@src/components/timeGrid/columnWith
 import { CurrentTimeIndicator } from '@src/components/timeGrid/currentTimeIndicator';
 import { MultipleTimezones } from '@src/components/timeGrid/multipleTimezones';
 import { getTopPercentByTime } from '@src/controller/times';
-import { TimeUnit, TimezoneConfig } from '@src/model';
 import EventUIModel from '@src/model/eventUIModel';
 import TZDate from '@src/time/date';
 import {
@@ -26,6 +25,9 @@ import {
   toStartOfDay,
 } from '@src/time/datetime';
 import { cls, toPercent, toPx } from '@src/util/cssHelper';
+
+import { TimeUnit } from '@t/events';
+import { TimezoneConfig } from '@t/option';
 
 const REFRESH_INTERVAL = 1000 * SIXTY_SECONDS;
