+++ apps/calendar/src/components/view/main.tsx
@@ -5,8 +5,9 @@ import { Day } from '@src/components/view/day';
 import { Month } from '@src/components/view/month';
 import { Week } from '@src/components/view/week';
 import { useStore } from '@src/contexts/calendarStore';
-import { ViewType } from '@src/model';
 import { viewSelector } from '@src/selectors';
+
+import { ViewType } from '@t/option';
 
 const views: {
   [k in ViewType]: FunctionComponent;
