+++ apps/calendar/src/components/view/month.tsx
@@ -8,7 +8,6 @@ import Panel from '@src/components/panel';
 import { MONTH_DAY_NAME_HEIGHT } from '@src/constants/style';
 import { useStore } from '@src/contexts/calendarStore';
 import { useTheme } from '@src/contexts/theme';
-import { MonthOption, TemplateMonthDayName } from '@src/model';
 import { optionSelector } from '@src/selectors';
 import { getGridInfo, getMonthCalendar, isWeekend } from '@src/time/datetime';
 import { cls } from '@src/util/cssHelper';
@@ -16,7 +15,9 @@ import { capitalizeDayName } from '@src/util/dayName';
 import { getSize } from '@src/util/dom';
 import { createMousePositionDataGrabberMonth } from '@src/util/viewHelper';
 
+import { MonthOption } from '@t/option';
 import { CalendarStore } from '@t/store';
+import { TemplateMonthDayName } from '@t/template';
 
 function getDayNames(option: CalendarStore['option']) {
   const { daynames, workweek } = option.month as Required<MonthOption>;
