+++ apps/calendar/src/contexts/calendarStore.ts
@@ -1,6 +1,5 @@
 import { useCallback } from 'preact/hooks';
 
-import { Option } from '@src/model';
 import { createCalendarDispatchers, createCalendarSlice } from '@src/slices/calendar';
 import { createOptionDispatchers, createOptionSlice } from '@src/slices/options';
 import { createPopupDispatchers, createPopupSlice } from '@src/slices/popup';
@@ -14,6 +13,7 @@ import { createStoreContext } from '@src/store';
 import { createStore } from '@src/store/internal';
 import { pick } from '@src/util/utils';
 
+import { Option } from '@t/option';
 import { CalendarStore, Dispatchers } from '@t/store';
 
 export const initCalendarStore = (option: Option = {}) =>
