+++ apps/calendar/src/controller/base.ts
@@ -1,10 +1,12 @@
-import { CalendarData, CalendarInfo, EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import EventUIModel from '@src/model/eventUIModel';
 import TZDate from '@src/time/date';
 import { makeDateRange, MS_PER_DAY, toEndOfDay, toFormat, toStartOfDay } from '@src/time/datetime';
 import Collection from '@src/util/collection';
 import { isSameEvent } from '@src/util/events';
+
+import { CalendarData, EventModelData } from '@t/events';
+import { CalendarInfo } from '@t/option';
 
 export type IDS_OF_DAY = Record<string, number[]>;
