+++ apps/calendar/src/controller/month.ts
@@ -9,13 +9,14 @@ import {
   limitRenderRange,
   positionUIModels,
 } from '@src/controller/core';
-import { CalendarData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import EventUIModel from '@src/model/eventUIModel';
 import TZDate from '@src/time/date';
 import { isSameDate, toEndOfDay, toFormat, toStartOfDay } from '@src/time/datetime';
 import array from '@src/util/array';
 import Collection, { Filter } from '@src/util/collection';
+
+import { CalendarData } from '@t/events';
 
 /**
  * Filter function for find allday event
