+++ apps/calendar/src/slices/view.ts
@@ -1,6 +1,6 @@
-import { ViewType } from '@src/model';
 import { deepMergedCopy } from '@src/util/utils';
 
+import { ViewType } from '@t/option';
 import { CalendarStore, SetState } from '@t/store';
 
 export type ViewSlice = {
