+++ apps/calendar/stories/columns.stories.tsx
@@ -4,7 +4,6 @@ import range from 'tui-code-snippet/array/range';
 
 import { GridSelectionInfo } from '@src/components/timeGrid';
 import { Column } from '@src/components/timeGrid/column';
-import { EventModelData } from '@src/model';
 import TZDate from '@src/time/date';
 import {
   addHours,
@@ -15,6 +14,8 @@ import {
   toStartOfYear,
 } from '@src/time/datetime';
 import { cls } from '@src/util/cssHelper';
+
+import { EventModelData } from '@t/events';
 
 import { createEventModels } from '@stories/helper/event';
