+++ apps/calendar/stories/helper/event.ts
@@ -1,6 +1,7 @@
-import { EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import EventUIModel from '@src/model/eventUIModel';
+
+import { EventModelData } from '@t/events';
 
 export function createEventModels(data: EventModelData[]): EventUIModel[] {
   return data.map((datum) => EventUIModel.create(EventModel.create(datum)));
