+++ apps/calendar/stories/monthView.stories.tsx
@@ -1,11 +1,11 @@
 import { h } from 'preact';
 
-import FloatingLayer from '@src/components/floatingLayer';
 import { Month } from '@src/components/view/month';
-import { EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
 import { addDate } from '@src/time/datetime';
+
+import { EventModelData } from '@t/events';
 
 import { ProviderWrapper } from '@stories/util/providerWrapper';
 import { createRandomEventModelsForMonth } from '@stories/util/randomEvents';
@@ -51,7 +51,6 @@ function createMonthEvents() {
 const Template: Story = (args) => (
   <ProviderWrapper options={args.options} events={args.events}>
     <Month />
-    <FloatingLayer />
   </ProviderWrapper>
 );
