+++ apps/calendar/test/app/controller/base.spec.ts
@@ -6,9 +6,10 @@ import {
   getDateRange,
   updateEvent,
 } from '@src/controller/base';
-import { CalendarData, EventModelData } from '@src/model';
 import EventModel from '@src/model/eventModel';
 import TZDate from '@src/time/date';
+
+import { CalendarData, EventModelData } from '@t/events';
 
 describe('controller/base', () => {
   let calendarData: CalendarData;
