+++ apps/calendar/test/app/controller/times.spec.ts
@@ -4,10 +4,10 @@ import {
   getPrevGridTimeFromMouseEvent,
   getTopPercentByTime,
 } from '@src/controller/times';
-import { TimeUnit } from '@src/model';
 import TZDate from '@src/time/date';
 import { cls } from '@src/util/cssHelper';
 
+import { TimeUnit } from '@t/events';
 import { createMouseEvent } from '@test/helper';
 
 interface TestData {
