+++ apps/calendar/types/grid.d.ts
@@ -1,0 +1,5 @@
+export interface GridUIModel {
+  day: number;
+  width: number;
+  left: number;
+}
