+++ apps/calendar/types/template.d.ts
@@ -1,68 +1,10 @@
 import { VNode } from 'preact';
 
-import { IDS_OF_DAY } from '@src/controller/base';
-import { EventHandler } from '@src/event';
-import { ExternalEventName } from '@src/event/externalEventType';
-import { InternalEventName } from '@src/event/internalEventType';
 import EventModel, { EventCategory } from '@src/model/eventModel';
-import { ThemeKeyValue } from '@src/theme/themeProps';
 import TZDate from '@src/time/date';
-import Collection from '@src/util/collection';
 
-import { TuiDateConstructor } from '@toast-ui/date';
-
-export type DateType = Date | string | number | TZDate;
-
-export interface EventModelData {
-  id?: string;
-  calendarId?: string;
-  title?: string;
-  body?: string;
-  start?: DateType;
-  end?: DateType;
-  goingDuration?: number;
-  comingDuration?: number;
-  isAllDay?: boolean;
-  category?: EventCategory;
-  dueDateClass?: string;
-  location?: string;
-  attendees?: string[];
-  recurrenceRule?: string;
-  isPending?: boolean;
-  isFocused?: boolean;
-  isVisible?: boolean;
-  isReadOnly?: boolean;
-  isPrivate?: boolean;
-  color?: string;
-  bgColor?: string;
-  dragBgColor?: string;
-  borderColor?: string;
-  customStyle?: string;
-  state?: string;
-  raw?: any;
-}
-
-export interface CalendarData {
-  calendars: CalendarInfo[];
-  events: Collection<EventModel>;
-  idsOfDay: IDS_OF_DAY;
-}
-
-export interface AppContext {
-  options: Option;
-  calendarData: CalendarData;
-  templates: Template;
-  internalEvent: EventHandler<InternalEventName>;
-  externalEvent: EventHandler<ExternalEventName>;
-}
-
-export interface GridUIModel {
-  day: number;
-  width: number;
-  left: number;
-}
-
-export type TimeUnit = 'second' | 'minute' | 'hour' | 'date' | 'month' | 'year';
+import { TimeUnit } from '@t/events';
+import { TimezoneConfig } from '@t/option';
 
 export interface TemplateTimeGridHourLabel {
   hidden: boolean;
@@ -157,83 +99,3 @@ export interface Template {
 }
 
 export type TemplateConfig = Partial<Template>;
-
-export interface WeekOption {
-  startDayOfWeek?: number;
-  daynames?: string[];
-  narrowWeekend?: boolean;
-  workweek?: boolean;
-  showTimezoneCollapseButton?: boolean;
-  timezonesCollapsed?: boolean;
-  timezones?: TimezoneConfig[];
-  hourStart?: number;
-  hourEnd?: number;
-}
-
-export interface MonthOption {
-  daynames?: string[];
-  startDayOfWeek?: number;
-  narrowWeekend?: boolean;
-  visibleWeeksCount?: number;
-  isAlways6Week?: boolean;
-  workweek?: boolean;
-  visibleEventCount?: number;
-  moreLayerSize?: {
-    width?: string | null;
-    height?: string | null;
-  };
-  grid?: {
-    header?: {
-      height?: number;
-    };
-    footer?: {
-      height?: number;
-    };
-  };
-  eventFilter?: (event: Required<EventModelData>) => boolean;
-}
-
-export interface CustomTimezone {
-  dateConstructor?: TuiDateConstructor; // YourCustomDate or LocalDate, UTCDate, MomentDate from @toast-ui/date;
-  offset?: number; // If using YourCustomDate or MomentDate
-  name?: string; // If using YourCustomDate or MomentDate
-}
-
-export interface TimezoneConfig {
-  timezoneOffset?: number;
-  timezoneName?: string;
-  displayLabel?: string;
-  tooltip?: string;
-}
-
-export interface CalendarColor {
-  color?: string;
-  bgColor?: string;
-  dragBgColor?: string;
-  borderColor?: string;
-}
-
-export interface CalendarInfo extends CalendarColor {
-  id: string;
-  name: string;
-}
-
-export type ViewType = 'month' | 'week' | 'day';
-
-export interface Option {
-  defaultView?: ViewType;
-  taskView?: boolean | string[];
-  eventView?: boolean | string[];
-  theme?: ThemeKeyValue;
-  template?: TemplateConfig;
-  week?: WeekOption;
-  month?: MonthOption;
-  calendars?: CalendarInfo[];
-  useCreationPopup?: boolean;
-  useDetailPopup?: boolean;
-  disableDblClick?: boolean;
-  disableClick?: boolean;
-  isReadOnly?: boolean;
-  usageStatistics?: boolean;
-  timezone?: CustomTimezone;
-}
